# Faster AVX2 dot product for IQ2_XS