There is a proposal in #5152 to change the bit arrangement for `IQ2_XS` quantization to speedup inference on `x86` CPUs. The proposed change improves performance by up 35% on CPU's with few cores, but is not backwards compatible and will require to redo quantization for all published `IQ2_XS`-quantized models.

To avoid breaking backwards compatibility, I decided to see by how much one can improve `AVX2` inference speed using the existing `IQ2_XS` data format, and this PR is the result. The table shows a comparison between master, the approach proposed in #5152, and this PR for TG-128 on a Ryzen-7950X CPU for Mistral-7B.

| model           | backend    |    threads | test       |      t/s (Master)|     t/s (#5152)  |       t/s (PR)   |   
| ----------------| ---------- | ---------: | ---------- | ---------------: | ---------------: | ---------------: |
| llama 7B IQ2_XS | CPU        |          2 | tg 128     |      4.78 ± 0.01 |      6.12 ± 0.01 |      5.71 ± 0.01 |
| llama 7B IQ2_XS | CPU        |          4 | tg 128     |      9.09 ± 0.02 |     11.48 ± 0.04 |     10.83 ± 0.02 |
| llama 7B IQ2_XS | CPU        |          8 | tg 128     |     16.31 ± 0.04 |     17.90 ± 0.02 |     19.04 ± 0.05 |
| llama 7B IQ2_XS | CPU        |         16 | tg 128     |     21.84 ± 0.54 |     22.02 ± 0.46 |     22.44 ± 0.08 |

The #5152 approach is still faster for 2 threads, but now the difference is only 7%. Strangely enough, the PR outperforms #5152 for 8 or more threads (I have observed such non-intuitive behavior in the past when working on CPU inference optimization for k-quants, where one approach may be faster for low number of threads but slower when using all threads).