@PeterReid 2024-01-29
Here are the timings on the CPUs I times #5152 on, as requested.

| CPU | t/s (master) | t/s (#5152) | t/s (PR)
-----|------|------|-----
| i5-4300U | 1.30 | 1.62 | 1.51
|i7-1165G7 | 2.55 | 2.95 | 2.70

I wonder if this reason it is outperforming #5152 on your CPU in some cases is that it is processing four blocks at a time instead of two.