@ikawrakow 2024-01-29
@PeterReid Thank you for these results. 

> I wonder if this reason it is outperforming https://github.com/ggerganov/llama.cpp/discussions/5152 on your CPU in some cases is that it is processing four blocks at a time instead of two.

I think it is complicated. At 8 or 16 threads the threads are fighting for access to the limited memory bandwidth, and stalling as a result. Instructions have latencies. The compiler has reordered the instructions to keep the pipeline as occupied as possible. It is difficult to predict how this gets affected by the stalls waiting for data.

In any case, your approach does save a few instructions and is clearly way more elegant. I'm just not sure if the few percent performance difference is worth to break backwards compatibility.