@PeterReid 2024-01-29
I have found a marginal improvement to make to this PR: https://github.com/PeterReid/llama.cpp/commit/12b4c1355bed8d5c82b093bd62da3873c2d87b0b . This removes two AND instructions (unnecessary because bit 8 is already known to be 0 because of the shift by 9).