@PeterReid 2024-01-30
I found a way to drop one of those shuffle instructions for bit counting by moving the xor before: https://github.com/PeterReid/llama.cpp/commit/03a3db128f2e8409a60e859e58095532142dd1a5 . And I've squashed them all together here if that makes more sense: https://github.com/PeterReid/llama.cpp/commit/aacf9a303c15632ef9b599f0c0582e8ea8042729  . Or should I just make a separate PR? I haven't done all the timing, but this cuts the difference between this PR and #5152 by about half for the little timing I've done.

It seems like with this shuffle trick you are using, there is very little computation saved by doing the backwards-compatibility break in #5152. The only argument for it I could see is that there may be platforms that don't have some fast way of bitcounting and for which doing the lookup table approach is slow.