@ikawrakow 2024-01-30
@PeterReid This is great! I was clearly confused by mostly operating on 8-bit SIMD vectors, where one does need to apply a mask after `_mm256_srli_epi16`. Thanks for 

With your changes, I get quite a bit better results on my Ryzen-7950X:

| model           | backend    |    threads | test       |              t/s |
| ----------------| ---------- | ---------: | ---------- | ---------------: |
| llama 7B IQ2_XS | CPU        |          2 | tg 128     |      6.70 ± 0.01 |
| llama 7B IQ2_XS | CPU        |          4 | tg 128     |     12.29 ± 0.05 |
| llama 7B IQ2_XS | CPU        |          8 | tg 128     |     21.09 ± 0.10 |
| llama 7B IQ2_XS | CPU        |         16 | tg 128     |     22.97 ± 0.34 |

Can you push your changes to this branch?